pyproject.toml
[tool.poetry]
name = "101-ai"
version = "0.1.0"
description = ""
authors = ["Alex Velikanov <alex25368@gmail.com>"]
readme = "README.md"
packages = [{include = "101_ai"}]

[tool.poetry.scripts]
start = "101_ai:run"

[tool.poetry.dependencies]
python = "^3.10"
quart = {extras = ["dotenv"], version = "^0.18.3"}
langchain = "^0.0.131"
google-search-results = "^2.4.2"
openai = "^0.27.4"
fixieai = "^0.2.14"
psycopg2-binary = "^2.9.6"
uvicorn = "0.20"
uvloop = "^0.17.0"
py-cord = "^2.4.1"
discord-interactions = "^0.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
